no pinta el coche aunque debería de pintarlo.

+++ practica4/frenet.cpp
@@ -3,7 +3,7 @@
 #include <QtCore/QDebug>
 #include <math.h>
 
-QList<PV3f> Frenet::marco(const QList<PV3f> &listaPuntos, GLdouble a, GLdouble b, GLdouble c, GLdouble t, int mF[16])
+QList<PV3f> Frenet::marco(const QList<PV3f> &listaPuntos, GLdouble a, GLdouble b, GLdouble c, GLdouble t, GLdouble mF[16])
 {
     const PV3f pc((a - b) * cos(t) + c * cos(t * (a - b) / b), 0,
                   (a - b) * sin(t) - c * sin(t * (a - b) / b), PV3f::Punto);
@@ -16,22 +16,22 @@ QList<PV3f> Frenet::marco(const QList<PV3f> &listaPuntos, GLdouble a, GLdouble b
     mF[0] = vn.getX();
     mF[1] = vn.getY();
     mF[2] = vn.getZ();
-    mF[3] = 0;
+    mF[3] = 0.0;
 
     mF[4] = vb.getX();
     mF[5] = vb.getY();
     mF[6] = vb.getZ();
-    mF[7] = 0;
+    mF[7] = 0.0;
 
     mF[8]  = vt.getX();
     mF[9]  = vt.getY();
     mF[10] = vt.getZ();
-    mF[11] = 0;
+    mF[11] = 0.0;
 
     mF[12] = pc.getX();
     mF[13] = pc.getY();
     mF[14] = pc.getZ();
-    mF[15] = 1;
+    mF[15] = 1.0;
 
     return multMatriz(listaPuntos, vn, vb, vt, a, b, c, t);
 }
